requirements.txt
numpy>=1.21.0
pandas>=1.3.0
ta-lib>=0.4.24
alpaca-py>=0.8.0
python-dotenv>=0.19.0
pytest>=6.2.5
jupyter>=1.0.0
matplotlib>=3.4.3
seaborn>=0.11.2
flask>=2.0.0 